Rename to Fabric and add license headers to source files
