+++ build.gradle
@@ -17,13 +17,15 @@ plugins {
 
     id 'idea'
     id 'eclipse'
+
+    id 'net.minecrell.licenser' version '0.2.1'
 }
 apply plugin: 'chorus.gradle'
 
 sourceCompatibility = 1.8
 targetCompatibility = 1.8
 
-group = 'chorusmc'
+group = 'net.fabricmc'
 archivesBaseName = project.name.toLowerCase()
 version = minecraft_version + '-0.0.1'
 
@@ -65,6 +67,11 @@ dependencies {
 	compile files(".gradle/minecraft/${minecraft_version}-client-mapped.jar")
 }
 
+license {
+    header file('HEADER')
+    include '**/*.java'
+}
+
 jar {
     from '.gradle/minecraft/mappings/'
 
@@ -96,20 +103,20 @@ uploadArchives {
                     url = project.url
 
                     scm {
-                        url = 'https://github.com/ChorusMC/chorus-base'
-                        connection = 'scm:git:git@github.com:ChorusMC/chorus-base.git'
-                        developerConnection = 'scm:git:git@github.com:ChorusMC/chorus-base.git'
+                        url = 'https://github.com/FabricMC/fabric-base'
+                        connection = 'scm:git:git@github.com:FabricMC/fabric-base.git'
+                        developerConnection = 'scm:git:git@github.com:FabricMC/fabric-base.git'
                     }
 
                     issueManagement {
                         system = 'GitHub'
-                        url = 'https://github.com/ChorusMC/chorus-base/issues'
+                        url = 'https://github.com/FabricMC/fabric-base/issues'
                     }
 
                     licenses {
                         license {
                             name = 'Apache License 2.0'
-                            url = 'https://github.com/ChorusMC/chorus-base/blob/master/LICENSE'
+                            url = 'https://github.com/FabricMC/fabric-base/blob/master/LICENSE'
                             distribution = 'repo'
                         }
                     }
