+++ gradle.properties
@@ -1,7 +1,7 @@
 # Basic stuff
-name = chorus-base
-description = The base component to the Chorus Mod Loader
-url = https://github.com/ChorusMC/chorus-base
+name = fabric-base
+description = The base component to the Fabric Mod Loader
+url = https://github.com/FabricMC/fabric-base
 
 # Minecraft stuff
 minecraft_version = 16w32b
