--- src/main/java/chorusmc/base/mixin/common/server/MixinMinecraftServer.java
@@ -1,15 +1,0 @@
-package chorusmc.base.mixin.common.server;
-
-import net.minecraft.server.MinecraftServer;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Overwrite;
-
-@Mixin(value = MinecraftServer.class, remap = false)
-public abstract class MixinMinecraftServer {
-
-    @Overwrite
-    public String getServerModName() {
-        return "Chorus";
-    }
-
-}
