+++ src/main/java/net/fabricmc/base/launch/ChorusClientMain.java
@@ -1,4 +1,20 @@
-package chorusmc.base.launch;
+/*
+ * Copyright 2016 FabricMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package net.fabricmc.base.launch;
 
 import com.google.gson.GsonBuilder;
 import com.mojang.authlib.Agent;
@@ -49,7 +65,7 @@ public class ChorusClientMain {
         arguments.addArgument("tweakClass", ChorusClientTweaker.class.getName());
 
         if (!arguments.containsArgument("accessToken")) {
-            arguments.addArgument("accessToken", "ChorusMC");
+            arguments.addArgument("accessToken", "FabricMC");
         }
 
         Launch.main(arguments.getArguments());
