+++ src/main/java/net/fabricmc/base/launch/ChorusClientTweaker.java
@@ -1,4 +1,20 @@
-package chorusmc.base.launch;
+/*
+ * Copyright 2016 FabricMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package net.fabricmc.base.launch;
 
 import net.minecraft.client.main.Main;
 import net.minecraft.launchwrapper.ITweaker;
@@ -36,7 +52,7 @@ public class ChorusClientTweaker implements ITweaker {
             this.args.put("--assetsDir", assetsDir.getAbsolutePath());
         }
         if (!this.args.containsKey("--accessToken")) {
-            this.args.put("--accessToken", "ChorusMC");
+            this.args.put("--accessToken", "FabricMC");
         }
         for (int i = 0; i < args.size(); i++) {
             String arg = args.get(i);
@@ -46,8 +62,8 @@ public class ChorusClientTweaker implements ITweaker {
         }
         MixinBootstrap.init();
         Mixins.addConfigurations(
-                "chorusmc.mixins.client.json",
-                "chorusmc.mixins.common.json");
+                "fabricmc.mixins.client.json",
+                "fabricmc.mixins.common.json");
         MixinEnvironment.getDefaultEnvironment().setSide(MixinEnvironment.Side.CLIENT);
     }
