+++ src/main/java/net/fabricmc/base/mixin/client/MixinClientBrandRetriever.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2016 FabricMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package net.fabricmc.base.mixin.client;
+
+import net.minecraft.client.ClientBrandRetriever;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Overwrite;
+
+@Mixin(value = ClientBrandRetriever.class, remap = false)
+public abstract class MixinClientBrandRetriever {
+
+    @Overwrite
+    public static String getClientModName() {
+        return "Fabric";
+    }
+
+}
