+++ src/main/java/net/fabricmc/base/mixin/client/gui/impl/MixinGuiMain.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2016 FabricMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package net.fabricmc.base.mixin.client.gui.impl;
+
+import net.minecraft.client.gui.GuiScreen;
+import net.minecraft.client.gui.impl.GuiMainMenu;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(value = GuiMainMenu.class, remap = false)
+public abstract class MixinGuiMain extends GuiScreen {
+
+    @Inject(method = "drawScreen(IIF)V", at = @At("RETURN"))
+    public void draw(int a1, int a2, float a3, CallbackInfo info) {
+        this.fontRenderer.drawString("Fabric Base", 2, this.height - 20, -1);
+    }
+
+}
