+++ src/main/resources/fabricmc.mixins.client.json
@@ -1,9 +1,10 @@
 {
   "required": true,
-  "package": "chorusmc.base.mixin.common",
+  "package": "net.fabricmc.base.mixin.client",
   "compatibilityLevel": "JAVA_8",
   "mixins": [
-    "server.MixinMinecraftServer"
+    "gui.impl.MixinGuiMain",
+    "MixinClientBrandRetriever"
   ],
   "injectors": {
     "defaultRequire": 1
